+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Added
+- `update` to `ContractSystem` to change config, current date or block lt.
+- `overwrite` to `ContractSystem` and `ContractExecutor` to replace contract state by address. Useful for mocking.
+
 ## [1.4.0] - 2023-01-08
 
 ## Added